Skip parsing if file doesn't contain assertions
Unassertify is quite an expensive transform, because it does full
parsing and code generation. That only needs to be done for files that
contain assertions, though. This patch adds a quick check: if the string
`assert` does not even occur in the file, we'll just skip it, saving
some time.